refactor(wizard): #1544 stage 4 — extract create_course reuseExistingCoursePath
Stage 4 of 8. Lifts the ~380-LOC reuse-existing-playbook branch from
create_course.ts (the L98-449 block in the pre-#1544 file) into two new
helpers:

  - tools/create_course/_reuse-path.ts (393 LOC) — the orchestration:
    dedup-by-name guard, #607 PlaybookSubject unlink invariant,
    PlaybookSource link (#352), COURSE_REFERENCE projection (#338),
    domain welcome message, test-caller enrollment, instant-curriculum
    kickoff, COURSE_REFERENCE source bridge, teachMethod backfill,
    pedagogy assertions (#1545), instruction-spec sync.
  - tools/create_course/_reuse-config-merge.ts (79 LOC) — pure function
    that builds the merged configUpdate from input + setupData with the
    #253 progressionMode and #167 pedagogy carryover.

Also introduces the shared `ResolvedCreateCourseContext` interface in
new file `tools/create_course/_context.ts` for Stages 4/5/7/8 to thread
domainId + subjectDiscipline + the orchestrator's input reads.

Behaviour-preserving. The dispatcher pin
`tests/lib/chat/wizard-tool-executor-dispatcher.test.ts` "create_course
(reuse path — #607 invariant)" still asserts the unlink call shape +
the updatePlaybookConfig reason string end-to-end through the
dispatcher → execute → helper chain.

Net file shape:
- create_course.ts: 1271 → 887 LOC (-384)
- new: _context.ts (31), _reuse-path.ts (393), _reuse-config-merge.ts (79)

Tests
- 11/11 dispatcher pins green (including #607 unlink + updatePlaybookConfig
  reason on the reuse path)
- 160/160 tests/lib/chat suite green
- 0 new tsc errors introduced
- AC respected: no file under tools/create_course/ exceeds 400 lines
  (largest is _reuse-path.ts at 393)

Refs #1544
